Document and find homes for functions
