+++ src/react/useMeldQuery.ts
@@ -1,41 +1,23 @@
 import { useEffect, useState } from "react";
-import { Observable, OperatorFunction, mergeMap } from "rxjs";
 
-import { type ReadQueryResult, readQuery } from "..";
+import { type ReadQueryResult } from "..";
+import { observeMeldQuery } from "../observable/observeMeldQuery";
 
-import type { MeldClone, MeldReadState, MeldUpdate } from "@m-ld/m-ld";
+import type { MeldClone } from "@m-ld/m-ld";
 import type { JsonValue } from "type-fest";
 
-const observeMeld = (meld: MeldClone) =>
-  new Observable<[MeldUpdate | null, MeldReadState]>((subscriber) => {
-    const subscription = meld.read(
-      async (state: MeldReadState) => {
-        subscriber.next([null, state]);
-      },
-      (update, state) =>
-        (async (state: MeldReadState) => {
-          subscriber.next([update, state]);
-        })(state)
-    );
-    return () => {
-      subscription.unsubscribe();
-    };
-  });
-
-const queryMap = <Data>(
-  query: JsonValue
-): OperatorFunction<
-  [MeldUpdate | null, MeldReadState],
-  ReadQueryResult<Data>
-> =>
-  mergeMap(
-    ([_update, state]) =>
-      // TODO: Replace this type assertion with actually derived types.
-      readQuery(state, query) as Promise<ReadQueryResult<Data>>
-  );
-
-// Use `Data` to explicitly assert the shape of the returned data.
-export const useMeldQuery = <Data>(
+/**
+ * A hook that subscribes to a Quild query over a m-ld clone and returns the
+ * latest result.
+ *
+ * @template Data The expected shape of the data returned by the query.
+ *                Eventually, this will be derived from the query itself. For
+ *                now, it must be given explicitly.
+ * @param meld The m-ld clone to query.
+ * @param query The Quild query to run.
+ * @returns The latest result of the query.
+ */
+export const useMeldQuery = <Data extends JsonValue>(
   meld: MeldClone | undefined,
   query: JsonValue
 ) => {
@@ -43,7 +25,7 @@ export const useMeldQuery = <Data>(
 
   useEffect(() => {
     if (meld) {
-      const o = observeMeld(meld).pipe(queryMap<Data>(query));
+      const o = observeMeldQuery<Data>(meld, query);
       const subscription = o.subscribe(setResult);
 
       return () => {
